requirements.txt
asgiref==3.2.7
astroid==2.3.3
click==7.1.2
colorama==0.4.3
Django==3.0.4
djangorestframework==3.11.0
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.2
opencv-python==4.2.0.32
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.11.2
